+++ src/components/Aapp/App.js
@@ -1,9 +1,9 @@
-import Header from '../header/Header'
-import Main from '../main/Main'
-import Footer from '../footer/Footer'
-import Home from '../home/Home'
+import Header from '../Aheader/Header'
+import Main from '../Amain/Main'
+import Footer from '../Afooter/Footer'
+import Home from '../Ahome/Home'
 import ActivitiesList from '../ActivitiesList/ActivitiesList'
-import CreateActivityForm from '../createActivityForm/CreateActivityForm'
+import CreateActivityForm from '../AcreateActivityForm/CreateActivityForm'
 import Map from '../Map/Map'
 import loadFromLocal from '../../lib/loadFromLocal'
 import saveToLocal from '../../lib/saveToLocal'
