+++ src/components/AflipCard/FlipCard.spec.js
