+++ src/components/Aheader/Header.js
@@ -1,6 +1,6 @@
-import BurgerButton from '../burgerButton/BurgerButton'
-import BurgerMenu from '../burgerMenu/BurgerMenu'
-import SearchBox from '../searchBox/SearchBox'
+import BurgerButton from '../AburgerButton/BurgerButton'
+import BurgerMenu from '../AburgerMenu/BurgerMenu'
+import SearchBox from '../AsearchBox/SearchBox'
 import logo from '../../assets/images/logo.svg'
 import styled from 'styled-components/macro'
 import { useState, useRef } from 'react'
