+++ src/components/AsearchBox/SearchBox.spec.js
